openSession is not compatible between Hive0.12 & 0.13.1
